+++ libs/designer/src/components/Main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useMemo } from "react";
 import { createEditorMachine } from "../machine";
 import { MachineContext } from "@paperclip-ui/common";
-import { DEFAULT_STATE } from "../machine/state";
+import { DEFAULT_STATE } from "../machine/state/core";
 import { env } from "../env";
 import { Editor } from "./Editor";
